Added nicer error message for alignment processed hmms
When people upload an HMM the website will set the alignment processing
option to convert to an HMM if it can. When it can't things will go
wrong as the default option is to use the observed counts processing.
Since we can detect the issue on the server side, we can change the
chosen value accordingly and give the results the user would be
expecting with a message explaining what happened.

There needed to be a different error message displayed on the site than
was being displayed via the API as the way to correct the issue was
different. This was done by checking to see if we think the request is
coming from a browser or some other software.

+++ lib/LogoServer/Controller/Root.pm
@@ -193,12 +193,19 @@ sub save_upload : Private {
     }
 
 
+
     if ($valid->{processing} =~ /^(?:weighted|observed)$/ && exists $c->stash->{file_is_hmm}) {
-      $c->stash->{error} = {
-        'processing' => 'HMM files do not require a post processing step. If you are uploading an hmm file, you can leave the processing argument blank.'
-      };
-      $c->stash->{rest}->{error} = $c->stash->{error};
-      $c->detach('end');
+      if ($c->request->looks_like_browser) {
+        $valid->{processing} = 'hmm';
+        $valid->{error} = 'HMM files do not require an alignment processing step. We have modified your alignment processing choice accordingly. In future please select the "Convert to an HMM" option to prevent this message.';
+      }
+      else {
+        $c->stash->{error} = {
+          'processing' => 'HMM files do not require an alignment processing step. If you are uploading an hmm file, you can leave the processing argument blank.'
+        };
+        $c->stash->{rest}->{error} = $c->stash->{error};
+        $c->detach('end');
+      }
     }
 
     # if we got nothing, then we don't want the json encode to blow up.
